requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap4==2024.10
cssselect2==0.8.0
Django==5.1.5
django-crispy-forms==2.4
django-filter==24.3
djangorestframework==3.15.2
fonttools==4.58.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
